+++ src/sockets.h
@@ -91,7 +91,7 @@ struct gnutella_socket {
 	struct getline *getline;	/* Line reader object */
 
 	gchar buffer[SOCK_BUFSZ];	/* buffer to put in the data read */
-	guint32 pos;				/* write position in the buffer */
+	size_t pos;			/* write position in the buffer */
 };
 
 /*
